requirements.txt
# Crypto Lake - Consolidated Requirements
# Base dependencies for production deployment

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database Support
duckdb>=0.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pg8000>=1.29.0                  # GCP Cloud SQL connector

# Configuration & Logging
pyyaml>=6.0
loguru>=0.7.0

# Networking
requests>=2.31.0
websockets>=11.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
google-auth>=2.0.0

# Machine Learning (QA)
scikit-learn>=1.3.0,<2.0.0

# Financial Data
yfinance>=0.2.0

# Cloud Storage
google-cloud-storage>=2.10.0

# Timezone Support
pytz>=2023.3

# Schema Validation (optional)
pandera>=0.17.0,<1.0.0

# Testing
pytest>=7.4.0
